Add Mode Selection -Ai Mode -Two Players
