+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TickTac"
         tools:targetApi="31">
+        <activity android:name=".MainActivity"
+            android:exported="false"/>
+        <activity
+            android:name=".TwoPlayerActivity"
+            android:exported="false" />
         <activity
             android:name=".SplashActivity"
             android:exported="true"
@@ -23,8 +28,8 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
-            android:exported="false"></activity>
+            android:name=".AiModeActivity"
+            android:exported="false" />
     </application>
 
 </manifest>
