+++ app/src/main/res/layout/activity_main.xml
@@ -1,70 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/fragmentContainerView"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-
-    android:gravity="center"
-    android:padding="16dp"
-
-    android:background="#203945"
-    >
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="TicTacToe With Ai "
-        android:textColor="@color/white"
-        android:textSize="20dp"
-        android:layout_marginBottom="40dp"/>
-
-    <Spinner
-        android:id="@+id/difficultySpinner"
-        android:layout_width="wrap_content"
-
-
-
-
-
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginBottom="30dp"
-        android:background="@drawable/button"
-        android:tooltipText="Select Game Level"
-
-
-
-        android:entries="@array/difficulty_levels"
-        android:textAlignment="center" />
-
-    <GridLayout
-        android:id="@+id/gridLayout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:columnCount="3"
-        android:rowCount="3">
-
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn00" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn01" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn02" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn10" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn11" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn12" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn20" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn21" style="@style/TicButton" android:background="@drawable/button"/>
-        <androidx.appcompat.widget.AppCompatButton android:id="@+id/btn22" style="@style/TicButton" android:background="@drawable/button"/>
-
-    </GridLayout>
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btnReset"
-        style="@style/Widget.AppCompat.Button.Borderless"
-        android:layout_width="300dp"
-        android:layout_height="wrap_content"
-        android:background="@drawable/button"
-        android:text="Reset"
-        android:textStyle="bold"
-        android:textColor="@color/white"
-
-        android:layout_marginTop="50dp" />
-
-</LinearLayout>
+    android:layout_height="match_parent"/>
