+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -1,12 +1,12 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent" android:layout_height="match_parent"
-    android:background="#121212">
+    android:background="@color/jaman">
 
     <ImageView
         android:id="@+id/logoImage"
         android:layout_width="150dp"
         android:layout_height="150dp"
-        android:src="@drawable/button"
+        android:src="@drawable/tictack"
         android:layout_centerHorizontal="true"
         android:layout_marginTop="80dp"/>
 
@@ -24,14 +24,40 @@
 
     <TextView
         android:id="@+id/versionInfo"
-        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Version 1.0\nDeveloped by Muhammad Uzair"
-        android:textSize="16sp"
+        android:layout_width="wrap_content"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentBottom="true"
+        android:layout_marginStart="60dp"
+        android:layout_marginEnd="60dp"
+        android:layout_marginBottom="40dp"
+        android:gravity="center"
+        android:textStyle="bold"
+        android:text="Developed by MU42"
         android:textColor="#ffffff"
-        android:layout_below="@id/appName"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="300dp"
-        android:gravity="center"/>
+        android:textSize="16sp" />
+
+
+        />
+
+    <TextView
+
+        android:layout_height="wrap_content"
+        android:layout_width="wrap_content"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentBottom="true"
+        android:layout_marginStart="60dp"
+        android:layout_marginEnd="60dp"
+        android:layout_marginBottom="20dp"
+        android:gravity="center"
+        android:textStyle="italic"
+        android:text="v1.2"
+        android:textColor="#ffffff"
+        android:textSize="10sp" />
+
+
+    />
 
 </RelativeLayout>
