+++ app/src/main/res/layout/fragment_mode_select.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:orientation="vertical"
+    android:background="@color/jaman"
+    android:padding="16dp">
+
+    <androidx.cardview.widget.CardView android:id="@+id/card2Player" android:layout_width="match_parent"
+        android:layout_height="150dp" android:layout_margin="16dp"
+        app:cardCornerRadius="20dp" app:cardElevation="10dp">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:gravity="center" android:orientation="horizontal">
+            <ImageView android:layout_width="60dp" android:layout_height="60dp" android:src="@drawable/button"/>
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:padding="20dp" android:text="2-Player Mode" android:textSize="22sp"/>
+        </LinearLayout>
+    </androidx.cardview.widget.CardView>
+
+    <androidx.cardview.widget.CardView android:id="@+id/cardAI" android:layout_width="match_parent"
+        android:layout_height="150dp" android:layout_margin="16dp"
+        app:cardCornerRadius="20dp" app:cardElevation="10dp">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:gravity="center" android:orientation="horizontal">
+            <ImageView android:layout_width="60dp"
+                android:layout_height="60dp" android:src="@drawable/button"/>
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:padding="20dp" android:text="AI Mode" android:textSize="22sp"/>
+        </LinearLayout>
+    </androidx.cardview.widget.CardView>
+
+</LinearLayout>
