+++ app/src/main/res/values/colors.xml
@@ -4,4 +4,5 @@
     <color name="white">#FFFFFFFF</color>
     <color name="aqua">#246660</color>
     <color name="dark_aqua">#053530</color>
+    <color name="jaman">#203945</color>
 </resources>
